vars

+++ konfuzio_sdk/data.py
@@ -187,7 +187,9 @@ class Page(Data):
 
     def get_image(self, update: bool = False) -> Image:
         """Get Document Page as PNG."""
-        if self.document.status[0] == 2 and (not is_file(self.image_path, raise_exception=False) or update):
+        if (self.document.status and self.document.status[0] == 2) and (
+            not is_file(self.image_path, raise_exception=False) or update
+        ):
             page_id = self.id_ if self.id_ else self.copy_of_id
             png_content = get_page_image(page_id)
             with open(self.image_path, "wb") as f:
